parsed the whole code to objects
